backend/requirements.txt
fastapi
uvicorn
pymilvus
sentence-transformers
google-generativeai
python-dotenv
